[#2] Refactor : 레이아웃 및 코드 정리
22.03.09 일 회의

- 불필요한 코드정리
- 폴더 추가 및 수정
- react import 삭제시 발생되는 lint 에러 "plugin:react/jsx-runtime" 규칙 추가
