+++ .eslintrc.json
@@ -9,8 +9,8 @@
 		"airbnb-typescript",
 		"airbnb/hooks",
 		"eslint:recommended",
-		"plugin:react/jsx-runtime",
 		"plugin:react/recommended",
+		"plugin:react/jsx-runtime",
 		"plugin:@typescript-eslint/recommended",
 		"prettier"
 	],
