+++ .prettierrc.json
@@ -7,3 +7,4 @@
 	"tabWidth": 2,
 	"overrides": [{ "files": ["*.json"], "options": { "parser": "json" } }]
 }
+
