+++ src/components/layout/Footer.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Box } from '@chakra-ui/react';
 
 function Footer() {
