+++ src/components/layout/Header.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Flex, Heading } from '@chakra-ui/react';
 import { Link } from 'react-router-dom';
