+++ src/components/layout/Layout.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Box } from '@chakra-ui/react';
 import Header from './Header';
 import Footer from './Footer';
